+++ packages/jaspr_cli/lib/src/dev/client_workflow.dart
@@ -44,15 +44,9 @@ class ClientWorkflow {
           return null;
         }
 
-        client.registerBuildTarget(
-          DefaultBuildTarget(
-            (b) => b
-              ..target = 'web'
-              ..reportChangedAssets = true,
-          ),
-        );
-
         logger.write('Starting initial build...', tag: Tag.builder, progress: ProgressState.running);
+
+        client.registerBuildTarget(DefaultBuildTarget((b) => b..target = 'web'));
         client.startBuild();
 
         final devProxy = await DevProxy.start(