+++ packages/jaspr_cli/lib/src/dev/dev_proxy.dart
@@ -41,7 +41,6 @@ class DevProxy {
     this.logger,
   }) {
     _listenToClientConnections();
-    _listenToBuildResults();
   }
 
   Future<void> _listenToClientConnections() async {
@@ -55,18 +54,34 @@ class DevProxy {
 
       _clientConnections[appId] = ClientConnection(appConnection, dwds, this)..start();
     }
+    _listenToBuildResults();
   }
 
   void _listenToBuildResults() async {
-    if (reload == ReloadConfiguration.hotReload) {
-      await for (final buildResult in buildResults) {
-        if (buildResult.status == BuildStatus.succeeded) {
+    await for (final buildResult in buildResults) {
+      if (buildResult.status == BuildStatus.succeeded) {
+        if (reload == ReloadConfiguration.hotReload) {
           for (final clientConnection in _clientConnections.values) {
             await clientConnection.performHotReload();
           }
         }
+
+        for (final callback in _postReloadCallbacks) {
+          // ignore: avoid_dynamic_calls
+          callback(buildResult);
+        }
       }
     }
+  }
+
+  final List<void Function(BuildResult result)> _postReloadCallbacks = [];
+
+  void registerPostReloadCallback(void Function(BuildResult result) callback) {
+    _postReloadCallbacks.add(callback);
+  }
+
+  void unregisterPostReloadCallback(void Function(BuildResult result) callback) {
+    _postReloadCallbacks.remove(callback);
   }
 
   ClientConnection? getClientConnection(String? appId) {
@@ -104,18 +119,17 @@ class DevProxy {
           }
         });
       }
-      final result = results.results.firstWhere((result) => result.target == target);
-      switch (result.status) {
-        case daemon.BuildStatus.started:
-          return BuildResult(status: BuildStatus.started);
-        case daemon.BuildStatus.failed:
-          return BuildResult(status: BuildStatus.failed);
-        case daemon.BuildStatus.succeeded:
-          return BuildResult(status: BuildStatus.succeeded);
-        default:
-          break;
+      final resultForTarget = results.results.where((result) => result.target == target).firstOrNull;
+      final result = switch (resultForTarget?.status) {
+        daemon.BuildStatus.started => BuildResult(status: BuildStatus.started),
+        daemon.BuildStatus.failed => BuildResult(status: BuildStatus.failed),
+        daemon.BuildStatus.succeeded => BuildResult(status: BuildStatus.succeeded),
+        _ => null,
+      };
+      if (result == null) {
+        throw StateError('Unexpected Daemon build result: $resultForTarget');
       }
-      throw StateError('Unexpected Daemon build result: $result');
+      return result;
     });
 
     var cascade = Cascade();