+++ packages/jaspr_cli/test/src/commands/serve_command_test.dart
@@ -18,7 +18,7 @@ void main() {
 
     setUp(() {
       io = FakeIO();
-      runner = JasprCommandRunner(false);
+      runner = JasprCommandRunner(false, false);
     });
 
     tearDown(() {
@@ -36,9 +36,9 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder([
-            'Running jaspr in client rendering mode.',
-            '[CLI] Starting web compiler...',
+          emitsInOrderWithTimeout([
+            'Starting myapp in client rendering mode.',
+            '[BUILDER] Starting web compilers...',
             '[BUILDER] Connecting to the build daemon...',
             '[BUILDER] Starting initial build...',
           ]),
@@ -53,15 +53,21 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder([
-            '[CLI] Done building web assets.',
-            '[CLI] Serving at http://localhost:8080',
+          emitsInOrderWithTimeout([
+            '[BUILDER] Done building web assets.',
+            '[LOG] Serving at http://localhost:8080',
           ]),
         );
 
         await io.shutdownBuildDaemon(buildDaemon);
 
-        expect(io.stdout.queue, emits('Terminating web compilers...'));
+        expect(
+          io.stdout.queue,
+          emitsInOrderWithTimeout([
+            'Stopping web compilers...',
+            'Terminating CSS runner...',
+          ]),
+        );
 
         expect(await serveResult, equals(0));
       });
@@ -79,10 +85,9 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder([
-            'Running jaspr in server rendering mode.',
-            'Using server entry point: lib/main.server.dart',
-            '[CLI] Starting web compiler...',
+          emitsInOrderWithTimeout([
+            'Starting myapp in server rendering mode.',
+            '[BUILDER] Starting web compilers...',
             '[BUILDER] Connecting to the build daemon...',
             '[BUILDER] Starting initial build...',
           ]),
@@ -90,7 +95,7 @@ void main() {
 
         await io.runInitialBuild(buildDaemon);
 
-        expect(io.stdout.queue, emits('[CLI] Done building web assets.'));
+        expect(io.stdout.queue, emits('[BUILDER] Done building web assets.'));
 
         await expectLater(
           io.serverSockets.next,
@@ -101,7 +106,13 @@ void main() {
 
         server.exit(0);
 
-        expect(io.stdout.queue, emits('Terminating web compilers...'));
+        expect(
+          io.stdout.queue,
+          emitsInOrderWithTimeout([
+            'Stopping web compilers...',
+            'Terminating CSS runner...',
+          ]),
+        );
 
         expect(await serveResult, equals(0));
       });
@@ -136,10 +147,9 @@ void main() {
 
         await expectLater(
           io.stdout.queue,
-          emitsInOrder([
-            'Running jaspr in server rendering mode.',
-            'Using server entry point: lib/main.server.dart',
-            '[CLI] Starting web compiler...',
+          emitsInOrderWithTimeout([
+            'Starting myapp in server rendering mode.',
+            '[BUILDER] Starting web compilers...',
             '[BUILDER] Connecting to the build daemon...',
             '[BUILDER] Starting initial build...',
           ]),
@@ -147,7 +157,7 @@ void main() {
 
         await io.runInitialBuild(buildDaemon);
 
-        expect(io.stdout.queue, emits('[CLI] Done building web assets.'));
+        expect(io.stdout.queue, emits('[BUILDER] Done building web assets.'));
 
         await io.connectToProxy();
 
@@ -159,8 +169,9 @@ void main() {
 
         expect(
           io.stdout.queue,
-          emitsInOrder([
-            'Terminating web compilers...',
+          emitsInOrderWithTimeout([
+            'Stopping web compilers...',
+            'Terminating CSS runner...',
             'Terminating flutter run...',
           ]),
         );
@@ -246,12 +257,18 @@ extension FakeServerIO on FakeIO {
   }
 
   Future<void> expectServerStarted(FakeProcess server) async {
-    await expectLater(this.stdout.queue, emits('[CLI] Starting server...'));
+    await expectLater(
+      this.stdout.queue,
+      emitsInOrderWithTimeout([
+        '[SERVER] Starting server...',
+        '[SERVER] Using server entry point: lib/main.server.dart',
+      ]),
+    );
 
     expect(fs.file('.dart_tool/jaspr/server_target.dart').existsSync(), isTrue);
     expect(fs.file('.dart_tool/jaspr/server.pid').existsSync(), isTrue);
 
-    await expectLater(this.stdout.queue, emits('[CLI] Server started.'));
+    await expectLater(this.stdout.queue, emits('[SERVER] Server started and listening on http://localhost:8080'));
 
     server.writeStdout('Fake server running.');