+++ packages/jaspr_router/lib/src/router.dart
@@ -52,6 +52,7 @@ class RouterState extends State<Router> with PreloadStateMixin {
   RouteMatchList get matchList => _matchList ?? RouteMatchList.empty;
 
   Map<Object, RouteLoader> routeLoaders = {};
+  VoidCallback? _routeChangeDispose;
 
   @override
   Future<void> preloadState() async {
@@ -64,7 +65,7 @@ class RouterState extends State<Router> with PreloadStateMixin {
   @override
   void initState() {
     super.initState();
-    PlatformRouter.instance.history.init(
+    _routeChangeDispose = PlatformRouter.instance.history.init(
       context,
       onChangeState: (state, {url}) {
         _update(url ?? context.url, extra: state, updateHistory: false, replace: true);
@@ -221,6 +222,13 @@ class RouterState extends State<Router> with PreloadStateMixin {
     final basePath = context.binding.basePath;
     final prefix = basePath.endsWith('/') ? basePath.substring(0, basePath.length - 1) : basePath;
     return prefix + location;
+  }
+
+  @override
+  void dispose() {
+    _routeChangeDispose?.call();
+    _routeChangeDispose = null;
+    super.dispose();
   }
 
   @override